% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exTREEmaTIME_SetAutoRates.R
\name{SetAutoRates}
\alias{SetAutoRates}
\title{estimates plausible values for rmin and rmax from input tree. Two alternative methods, 1) based on sister pairs of terminal branches, 2) based on variance of root to tip distance
outputs rmax_calc and rmin_calc for full analysis in exTREEmaTIME}
\usage{
SetAutoRates(
  tree,
  auto_rates_type,
  max_rate_unit,
  min_rate_unit,
  noise_level,
  root_min,
  root_max
)
}
\description{
estimates plausible values for rmin and rmax from input tree. Two alternative methods, 1) based on sister pairs of terminal branches, 2) based on variance of root to tip distance
outputs rmax_calc and rmin_calc for full analysis in exTREEmaTIME
}
